fix(php85): casts no canónicos (double) -> (float)
Fase D, familia 1 de 3. Nueve ocurrencias en siete archivos.

PHP 8.5 deprecó la grafía no canónica del cast. (double) y (float) son el
mismo cast, así que el cambio no altera la semántica: solo la escritura.

Importa porque la deprecación se emite en TIEMPO DE COMPILACIÓN, al cargar
el archivo, sin ejecutar la función que la contiene. Con la promoción de
E_DEPRECATED a excepción de bootstrap.php, bastaba con que el autoloader
tocara cualquiera de estos siete archivos para matar la petición en 8.5.

Familia cerrada: no quedan (integer), (boolean), (binary) ni (real).

+++ src/app/classes/ApplicationCalls/Controllers/ApplicationCallsController.php
@@ -470,7 +470,7 @@ class ApplicationCallsController extends AdminPanelController
                 },
                 false,
                 function ($value) {
-                    return Validator::isDouble($value) ? (double) $value : $value;
+                    return Validator::isDouble($value) ? (float) $value : $value;
                 }
             ),
             new Parameter(
