+++ src/app/classes/Organizations/Controllers/OrganizationsController.php
@@ -478,7 +478,7 @@ class OrganizationsController extends AdminPanelController
                 },
                 true,
                 function ($value) {
-                    return Validator::isDouble($value) ? (double) $value : $value;
+                    return Validator::isDouble($value) ? (float) $value : $value;
                 }
             ),
             new Parameter(
@@ -489,7 +489,7 @@ class OrganizationsController extends AdminPanelController
                 },
                 true,
                 function ($value) {
-                    return Validator::isDouble($value) ? (double) $value : $value;
+                    return Validator::isDouble($value) ? (float) $value : $value;
                 }
             ),
             new Parameter(
