+++ src/app/core/psr4/PiecesPHP/Core/Sitemap/Sitemap.php
@@ -77,7 +77,7 @@ class Sitemap
                                 $changeFreq = (string) $tag;
                             }
                             if ($name == 'priority') {
-                                $priority = (double) ((string) $tag);
+                                $priority = (float) ((string) $tag);
                             }
 
                         }
